simd distance function for radius queries

+++ crates/analysis/src/incidence.rs
@@ -11,8 +11,8 @@ use crate::incidence_iter::{iter_incidences, IncidenceIter};
 
 /// An incidence view of a hypergraph.
 ///
-/// This provides an iterator over the `(node, edge, sign)` tuples of the
-/// hypergraph, where `sign` is `1` for outputs and `-1` for inputs.
+/// This provides an iterator over the `(node, edge, weight)` tuples of the
+/// hypergraph, where `weight` is `1.0` for outputs and `-1.0` for inputs.
 pub struct IncidenceView<'a> {
     hypergraph: &'a HypergraphFrozen,
 }
@@ -30,7 +30,7 @@ impl<'a> FrozenHypergraphView<'a> for IncidenceView<'a> {
 
     fn for_each<F>(&self, mut f: F) -> ControlFlow<()>
     where
-        F: FnMut((NodeIndex, EdgeIndex, i8)) -> ControlFlow<()>,
+        F: FnMut((NodeIndex, EdgeIndex, f64)) -> ControlFlow<()>,
     {
         for item in self.iter_view() {
             if f(item).is_break() {
