+++ crates/hypergraph/src/lib.rs
@@ -14,7 +14,7 @@ use std::ops::ControlFlow;
 pub use core::ids::{EdgeIx as EdgeIndex, NodeIx as NodeIndex};
 
 pub trait FrozenHypergraphView<'a> {
-    type Iter: Iterator<Item = (NodeIndex, EdgeIndex, i8)> + 'a;
+    type Iter: Iterator<Item = (NodeIndex, EdgeIndex, f64)> + 'a;
 
     fn new_view(hypergraph: &'a frozen::HypergraphFrozen) -> Self;
 
@@ -22,7 +22,7 @@ pub trait FrozenHypergraphView<'a> {
 
     fn for_each<F>(&self, f: F) -> ControlFlow<()>
     where
-        F: FnMut((NodeIndex, EdgeIndex, i8)) -> ControlFlow<()>;
+        F: FnMut((NodeIndex, EdgeIndex, f64)) -> ControlFlow<()>;
 }
 
 pub trait Vertex {}
